use batch bitstream

+++ ref/include/bitstream_batch.h
@@ -16,7 +16,7 @@ int bs_batch_init(bitstream_batch_t *bs, uint8_t **buf, size_t buf_len,
                   size_t batch_size);
 
 // int bs_batch_write(bitstream_batch_t *bs, uint16_t **data, uint32_t pos, uint32_t data_len);
-int bs_batch_write(bitstream_batch_t *bs, uint16_t *data, uint32_t data_len);
+int bs_batch_write(bitstream_batch_t *bs, uint32_t *data, uint32_t data_len);
 
 int bs_batch_read(bitstream_batch_t *bs, uint32_t *buf, uint32_t data_len);
